backend/requirements.txt
fastapi==0.115.12
pandas==2.2.3
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.0.1
pytz==2025.1
typing_extensions==4.12.2
uvicorn==2.1.1